ktp/leetcode-bot/requirements.txt
Flask==2.0.1
slack_sdk==3.15.0
python-dotenv==0.20.0